+++ CHANGES.md
@@ -1,3 +1,8 @@
+0.1.6 (2020-04-20)
+-------------------
+- Added a convienience function which wavelength calibrates from a list
+of spectral feature pixel and order positions, and a provided reference line list.
+
 0.1.5 (2020-03-22)
 -------------------
 - Feature identification stage now fetches the pixel error in the centroid, and the flux error